+++ src/zope/mimetype/tests/test_doctests.py
@@ -43,7 +43,7 @@ def optional_test_suite_setUp(required_module):
         import importlib
         try:
             importlib.import_module(required_module)
-        except ImportError:
+        except ImportError: # pragma: no cover
             raise unittest.SkipTest("Required module %r missing" % (required_module,))
         testing.setUp(test)
     return setUp