+++ src/zope/mimetype/utils.py
@@ -13,10 +13,11 @@
 """Utility helpers
 """
 try:
-    from email.Charset import Charset
-except ImportError:
-    # Py3: Changed package name.
     from email.charset import Charset
+except ImportError: # pragma: no cover
+    # Py2: Old name. (Note, on case-insensitive filesystems,
+    # we can import with either. This complicates coverage.)
+    from email.Charset import Charset
 
 
 def decode(s, charset_name):